Add files via upload
A bookmark app, in which you can add the name of your favorite site and can visit the site from the app. This app is helpful at the time when you visited a site long  before and have forgotten about it, at that time you can visit that site using this app.